Move to util function
